Add missing calls to popInstantiationLimit()
Prior to this change, popInstantiationLimit() was not performed
when a visible candidate is instantiated, then not selected
as a/the best candidate.

This change adds a call to popInstantiationLimit() to cover
this case. Also it adds a call to explainAndCheckInstantiation()
to cover a case when popInstantiationLimit() would be invoked
without prior incrementing of instantiationLimitMap.